+++ c++/UziFirmware/Script.h
@@ -18,7 +18,7 @@ struct Script
 	uint8* locals = 0;
 
 	int16 instructionStart = 0;
-	uint8 instructionCount = 0;
+	uint16 instructionCount = 0;
 	Instruction* instructions = 0;
 
 	Coroutine* coroutine = 0;
@@ -26,7 +26,7 @@ struct Script
 	uint8 getIndex(void);
 	int16 getInstructionStart(void);
 	int16 getInstructionStop(void);
-	uint8 getInstructionCount(void);
+	uint16 getInstructionCount(void);
 	Instruction getInstructionAt(int16);
 	bool isStepping(void);
 	void setStepping(bool);
@@ -41,4 +41,4 @@ struct Script
 	void setBreakpointAt(int16, bool);
 };
 
-Error readScript(Reader * rs, Script* script, uint8 start, uint8 scriptIndex, float* globals);
+Error readScript(Reader * rs, Script* script, int16 start, uint8 scriptIndex, float* globals);
